Move pidnet function
