+++ src/netspresso_trainer/dataloaders/augmentation/transforms.py
@@ -1,8 +1,31 @@
-from torchvision.transforms.functional import InterpolationMode
+import cv2
+import numpy as np
+import PIL.Image as Image
 
 from . import custom as TC
 from .registry import TRANSFORM_DICT
 from ..utils.constants import IMAGENET_DEFAULT_MEAN, IMAGENET_DEFAULT_STD
+
+EDGE_SIZE = 4
+Y_K_SIZE = 6
+X_K_SIZE = 6
+
+
+def reduce_label(label: np.ndarray) -> Image.Image:
+    label[label == 0] = 255
+    label = label - 1
+    label[label == 254] = 255
+    return Image.fromarray(label)
+
+
+def generate_edge(label: np.ndarray) -> Image.Image:
+    edge = cv2.Canny(label, 0.1, 0.2)
+    kernel = np.ones((EDGE_SIZE, EDGE_SIZE), np.uint8)
+    # edge_pad == True
+    edge = edge[Y_K_SIZE:-Y_K_SIZE, X_K_SIZE:-X_K_SIZE]
+    edge = np.pad(edge, ((Y_K_SIZE, Y_K_SIZE), (X_K_SIZE, X_K_SIZE)), mode='constant')
+    edge = (cv2.dilate(edge, kernel, iterations=1) > 50) * 1.0
+    return Image.fromarray((edge.copy() * 255).astype(np.uint8))
 
 
 def transforms_custom_train(conf_augmentation):
