+++ src/netspresso_trainer/dataloaders/segmentation/huggingface.py
@@ -4,7 +4,7 @@ import numpy as np
 import PIL.Image as Image
 
 from ..base import BaseHFDataset
-from ..segmentation.transforms import generate_edge, reduce_label
+from ..augmentation.transforms import generate_edge, reduce_label
 
 
 class SegmentationHFDataset(BaseHFDataset):
